backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
supabase==2.3.0
fastapi-cache2[redis]==0.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
authlib==1.3.0
stripe==8.0.0
